Correcao de erro apos merge com as alteracoes de refatoracoes feitas em outros cards que impactaram neste aqui. - Migracao de testes que vao na superpay para classe mockada. - Outros ajustes
