+++ src/main/java/com/cosmeticos/payment/superpay/client/rest/model/DadosCartao.java
@@ -2,11 +2,13 @@ package com.cosmeticos.payment.superpay.client.rest.model;
 
 import lombok.Data;
 
+import java.util.Date;
+
 @Data
 public class DadosCartao {
     private String nomePortador;
     private String numeroCartao;
     private String codigoSeguranca;
-    private String dataValidade;
+    private Date dataValidade;
     
 }
