+++ src/test/java/com/cosmeticos/controller/PaymentControllerTests.java
@@ -2,12 +2,8 @@ package com.cosmeticos.controller;
 
 import com.cosmeticos.Application;
 import com.cosmeticos.commons.CampainhaSuperpeyResponseBody;
-import com.cosmeticos.commons.CustomerResponseBody;
-import com.cosmeticos.commons.OrderResponseBody;
-import com.cosmeticos.model.*;
-import com.cosmeticos.repository.*;
-import com.cosmeticos.service.OrderService;
-import com.cosmeticos.service.VoteService;
+import com.cosmeticos.model.Category;
+import com.cosmeticos.repository.CategoryRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.junit.Assert;
 import org.junit.Before;
@@ -24,9 +20,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.sql.Timestamp;
 import java.text.ParseException;
-import java.time.LocalDateTime;
 
 /**
  * Created by matto on 08/08/2017.
@@ -39,295 +33,18 @@ public class PaymentControllerTests {
     private TestRestTemplate restTemplate;
 
     @Autowired
-    private CustomerRepository customerRepository;
-
-    @Autowired
-    private ServiceRepository serviceRepository;
-
-    @Autowired
-    private ProfessionalRepository professionalRepository;
-
-    @Autowired
-    private AddressRepository addressRepository;
-
-    @Autowired
-    private ScheduleRepository scheduleRepository;
-
-    @Autowired
-    private OrderRepository orderRepository;
-
-    @Autowired
-    private WalletRepository walletRepository;
-
-    @Autowired
-    private OrderService service;
-
-    @Autowired
-    private VoteService voteService;
-
-    @Autowired
-    private PaymentController paymentController;
+    private CategoryRepository categoryRepository;
 
     @Before
     public void setup()
     {
-        Service service = serviceRepository.findByCategory("PEDICURE");
+        Category category = categoryRepository.findByName("PEDICURE");
 
-        if(service == null) {
-            service = new Service();
-            service.setCategory("PEDICURE");
-            serviceRepository.save(service);
+        if(category == null) {
+            category = new Category();
+            category.setName("PEDICURE");
+            categoryRepository.save(category);
         }
-    }
-
-    @Test
-    public void testPaymentOk() throws URISyntaxException, ParseException, JsonProcessingException {
-
-        /*
-         PRE-CONDICOES para o teste:
-         Criamos um Customer qualquer.
-         Criamos um Profissional qualquer e o associamos a um Service.
-         Salvamos tudo no banco.
-         Criamos uma Order
-         */
-
-        //-------- INICIO DA CRIACAO DE CUSTOMER ----------/
-
-        String emailCustomer = "testPaymentOk-customer1@email.com";
-
-        String jsonCustomerCreate = "{\n" +
-                "   \"customer\":{\n" +
-                "      \"address\":{\n" +
-                "         \"address\": \"Avenida dos Metalúrgicos, 22\",\n" +
-                "         \"cep\":\"26083-275\",\n" +
-                "         \"city\":\"Nova Iguaçu\",\n" +
-                "         \"country\":\"BR\",\n" +
-                "         \"neighborhood\":\"Rodilândia\",\n" +
-                "         \"state\":\"RJ\"\n" +
-                "      },\n" +
-                "      \"birthDate\":1310353200000,\n" +
-                "      \"cellPhone\":null,\n" +
-                "      \"dateRegister\":null,\n" +
-                "      \"genre\":null,\n" +
-                "      \"status\":null,\n" +
-                "      \"user\":{\n" +
-                "         \"email\":\""+ emailCustomer +"\",\n" +
-                "         \"idLogin\":null,\n" +
-                "         \"password\":\"123\",\n" +
-                "         \"sourceApp\":null,\n" +
-                "         \"username\":\""+ emailCustomer +"\"\n" +
-                "      },\n" +
-                "      \"cpf\":\"123.605.789-05\",\n" +
-                "      \"idAddress\":null,\n" +
-                "      \"idCustomer\":null,\n" +
-                "      \"idLogin\":null,\n" +
-                "      \"nameCustomer\":\"testPaymentOk-customer1\"\n" +
-                "   }\n" +
-                "}";
-
-        System.out.println(jsonCustomerCreate);
-
-        RequestEntity<String> entityCustomer =  RequestEntity
-                .post(new URI("/customers"))
-                .contentType(MediaType.APPLICATION_JSON)
-                .accept(MediaType.APPLICATION_JSON)
-                .body(jsonCustomerCreate);
-
-        ResponseEntity<CustomerResponseBody> exchange = restTemplate
-                .exchange(entityCustomer, CustomerResponseBody.class);
-
-        Assert.assertNotNull(exchange);
-        Assert.assertEquals(HttpStatus.OK, exchange.getStatusCode());
-        Assert.assertEquals("testPaymentOk-customer1", exchange.getBody().getCustomerList().get(0).getNameCustomer());
-
-        //ABAIXO SEGUE O CUSTOMER QUE FOI RETORNADO APOS CRIAR ACIMA, NOTE QUE O ID DE ADDRESS RETORNADO FOI 15
-        Customer customer = exchange.getBody().getCustomerList().get(0);
-
-        //TODO - AO BUSCAR NO BANCO O CUSTOMER PELO ID, O ADDRESS RETORNADO NAO EH O MESMO QUE FOI CRIADO INICIALMENTE
-        //ABAIXO SEGUE O CUSTOMER QUE BUSCAMOS NO BANCO PELO ID DO CUSTOMER CRIADO ACIMA, O ID DE ADDRESS RETORNADO FOI 7
-        Customer customer2 = customerRepository.findOne(customer.getIdCustomer());
-
-        //-------- FIM DA CRIACAO DE CUSTOMER ----------/
-
-        Professional professional = ProfessionalControllerTests.createFakeProfessional();
-        professional.getUser().setUsername("testPaymentOk-professional");
-        professional.getUser().setEmail("testPaymentOk-professional@email.com");
-        professional.getUser().setPassword("123");
-        professional.setCnpj("123.605.789-06");
-
-        customerRepository.save(customer);
-        professionalRepository.save(professional);
-
-
-        Service service = serviceRepository.findByCategory("PEDICURE");
-
-        ProfessionalServices ps1 = new ProfessionalServices(professional, service);
-
-        professional.getProfessionalServicesCollection().add(ps1);
-
-        // Atualizando associando o Profeissional ao Servico
-        professionalRepository.save(professional);
-
-        /*
-        //JSON PARA CRIAR ORDER PARA EFETUAR O PAGAMENTO
-        String jsonCreateOrder = "{\n" +
-                "  \"order\" : {\n" +
-                "    \"date\" : 1498324200000,\n" +
-                "    \"status\" : 0,\n" +
-                "    \"scheduleId\" : {\n" +
-                "      \"scheduleDate\" : \""+ Timestamp.valueOf(LocalDateTime.MAX.of(2017, 07, 05, 12, 10, 0)).getTime() +"\",\n" +
-                "      \"status\" : \"ACTIVE\",\n" +
-                "      \"orderCollection\" : [ ]\n" +
-                "    },\n" +
-                "    \"professionalServices\" : {\n" +
-                "      \"service\" : {\n" +
-                "        \"idService\" : "+ service.getIdService() +",\n" +
-                "        \"category\" : \"PEDICURE\"\n" +
-                "      },\n" +
-                "      \"professional\" : {\n" +
-                "        \"idProfessional\" : "+ professional.getIdProfessional() +",\n" +
-                "        \"nameProfessional\" : \""+ professional.getNameProfessional() +"\",\n" +
-                "        \"cnpj\" : \""+ professional.getIdProfessional() +"\",\n" +
-                "        \"genre\" : \"F\",\n" +
-                "        \"birthDate\" : 688010400000,\n" +
-                "        \"cellPhone\" : \"(21) 99887-7665\",\n" +
-                "        \"dateRegister\" : 1499195092952,\n" +
-                "        \"status\" : 0\n" +
-                "      }\n" +
-                "    },\n" +
-                "    \"idLocation\" : null,\n" +
-                "    \"idCustomer\" : {\n" +
-                "      \"idCustomer\" : "+ customer.getIdCustomer() +",\n" +
-                "      \"nameCustomer\" : \""+ customer.getNameCustomer() +"\",\n" +
-                "      \"cpf\" : \""+ customer.getCpf() +"\",\n" +
-                "      \"genre\" : \"F\",\n" +
-                "      \"birthDate\" : 688010400000,\n" +
-                "      \"cellPhone\" : \"(21) 99887-7665\",\n" +
-                "      \"dateRegister\" : 1499195092952,\n" +
-                "      \"status\" : 0,\n" +
-                "      \"idLogin\" : {\n" +
-                "        \"username\" : \""+ customer.getUser().getUsername() +"\",\n" +
-                "        \"email\" : \""+ customer.getUser().getEmail() +"\",\n" +
-                "        \"password\" : \""+ customer.getUser().getPassword() +"\",\n" +
-                "        \"sourceApp\" : \"facebook\"\n" +
-                "      },\n" +
-                //"      \"idAddress\" : null\n" +
-                "       \"idAddress\": { \n" +
-                "   	    \"address\": \"Avenida dos Metalúrgicos, 22\",\n" +
-                "   	    \"cep\": \"26083-275\",\n" +
-                "   	    \"neighborhood\": \"Rodilândia\",\n" +
-                "   	    \"city\": \"Nova Iguaçu\",\n" +
-                "   	    \"state\": \"RJ\",\n" +
-                "   	    \"country\": \"BR\" \n" +
-                "       },\n" +
-                "       \"address\": { \n" +
-                "   	    \"address\": \"Avenida dos Metalúrgicos, 22\",\n" +
-                "   	    \"cep\": \"26083-275\",\n" +
-                "   	    \"neighborhood\": \"Rodilândia\",\n" +
-                "   	    \"city\": \"Nova Iguaçu\",\n" +
-                "   	    \"state\": \"RJ\",\n" +
-                "   	    \"country\": \"BR\" \n" +
-                "       }\n" +
-                "    }\n" +
-                "  }\n" +
-                "}";
-        */
-
-        //JSON PARA CRIAR ORDER PARA EFETUAR O PAGAMENTO
-        String jsonCreateOrder = "{\n" +
-                "  \"order\" : {\n" +
-                "    \"date\" : 1498324200000,\n" +
-                "    \"status\" : 0,\n" +
-                "    \"scheduleId\" : {\n" +
-                "      \"scheduleDate\" : \""+ Timestamp.valueOf(LocalDateTime.MAX.of(2017, 07, 05, 12, 10, 0)).getTime() +"\",\n" +
-                "      \"status\" : \"ACTIVE\",\n" +
-                "      \"orderCollection\" : [ ]\n" +
-                "    },\n" +
-                "    \"professionalServices\" : {\n" +
-                "      \"service\" : {\n" +
-                "        \"idService\" : "+ service.getIdService() +",\n" +
-                "        \"category\" : \"PEDICURE\"\n" +
-                "      },\n" +
-                "      \"professional\" : {\n" +
-                "        \"idProfessional\" : "+ professional.getIdProfessional() +",\n" +
-                "        \"nameProfessional\" : \""+ professional.getNameProfessional() +"\",\n" +
-                "        \"cnpj\" : \""+ professional.getIdProfessional() +"\",\n" +
-                "        \"genre\" : \"F\",\n" +
-                "        \"birthDate\" : 688010400000,\n" +
-                "        \"cellPhone\" : \"(21) 99887-7665\",\n" +
-                "        \"dateRegister\" : 1499195092952,\n" +
-                "        \"status\" : 0\n" +
-                "      }\n" +
-                "    },\n" +
-                "    \"idLocation\" : null,\n" +
-                "    \"idCustomer\" : {\n" +
-                "      \"idCustomer\" : "+ customer.getIdCustomer() +",\n" +
-                "      \"nameCustomer\" : \""+ customer.getNameCustomer() +"\",\n" +
-                "      \"cpf\" : \""+ customer.getCpf() +"\",\n" +
-                "      \"genre\" : \"F\",\n" +
-                "      \"birthDate\" : 688010400000,\n" +
-                "      \"cellPhone\" : \"(21) 99887-7665\",\n" +
-                "      \"dateRegister\" : 1499195092952,\n" +
-                "      \"status\" : 0,\n" +
-                "      \"idLogin\" : {\n" +
-                "        \"username\" : \""+ customer.getUser().getUsername() +"\",\n" +
-                "        \"email\" : \""+ customer.getUser().getEmail() +"\",\n" +
-                "        \"password\" : \""+ customer.getUser().getPassword() +"\",\n" +
-                "        \"sourceApp\" : \"facebook\"\n" +
-                "      },\n" +
-                "      \"idAddress\" : "+ customer.getAddress().getIdAddress() +",\n" +
-                "      \"address\": { \n" +
-                "   	    \"idAddress\": "+ customer.getAddress().getIdAddress() +"\n" +
-                "      }\n" +
-                "    }\n" +
-                "  }\n" +
-                "}";
-
-        System.out.println(jsonCreateOrder);
-
-        RequestEntity<String> entity =  RequestEntity
-                .post(new URI("/orders"))
-                .contentType(MediaType.APPLICATION_JSON)
-                .accept(MediaType.APPLICATION_JSON)
-                .body(jsonCreateOrder);
-
-        ResponseEntity<OrderResponseBody> exchangeCreate = restTemplate
-                .exchange(entity, OrderResponseBody.class);
-        //TODO - NAO SEI POR QUAL MOTIVO, MAS OS DADOS DO ENDERECO NAO ESTAO VINDO - PARECE QUE NAO ESTA SALVANDO EM ORDER CREATE
-        //-- checar se o seu teste esta incluindo address no json
-        Assert.assertNotNull(exchangeCreate);
-        Assert.assertNotNull(exchangeCreate.getBody().getOrderList());
-        Assert.assertEquals(HttpStatus.OK, exchangeCreate.getStatusCode());
-
-        Order order = exchangeCreate.getBody().getOrderList().get(0);
-        Order order1 = orderRepository.findOne(order.getIdOrder());
-
-        //TODO - OS DO ENDERECO NAO ESTAO VINDO
-        Address address = addressRepository.findOne(order1.getIdCustomer().getAddress().getIdAddress());
-
-        /************ FIM DAS PRE_CONDICOES **********************************/
-
-        String retornoSuperpay = paymentController.sendRequest(order);
-
-        Assert.assertNotNull(retornoSuperpay);
-        System.out.println(retornoSuperpay);
-
-    }
-
-    //TODO - COMO ESSE TESTE CHAMA A URL DA SUPERPAY, DEVEMOS COLOCAR EM UMA NOVA CLASSE DE TESTES MOCKADA
-    @Test
-    public void testCampainhaOK() throws URISyntaxException, ParseException, JsonProcessingException {
-
-        String numeroTransacao = "3";
-        String codigoEstabelecimento = "1501698887865";
-        String campoLivre1 = "TESTE";
-
-        ResponseEntity<CampainhaSuperpeyResponseBody> exchange = this.executaCampainha(
-                numeroTransacao, codigoEstabelecimento, campoLivre1);
-
-        Assert.assertNotNull(exchange);
-        Assert.assertEquals(HttpStatus.OK, exchange.getStatusCode());
-
     }
 
     @Test
@@ -359,9 +76,6 @@ public class PaymentControllerTests {
         Assert.assertEquals(HttpStatus.NOT_FOUND, exchange.getStatusCode());
 
     }
-
-
-
 
     public ResponseEntity<CampainhaSuperpeyResponseBody> executaCampainha(
             String numeroTransacao, String codigoEstabelecimento, String campoLivre1)
